componentsを単数型に
