+++ src/component/context/global/reducer.ts
