+++ src/component/pages/page1/confirm/page.tsx
@@ -1,22 +1,22 @@
 import type { NextPageWithLayout } from 'next';
 import Head from 'next/head';
-import Page1 from '@/components/pages/page1/page';
+import Page1 from '@/component/pages/page1/page';
 
 type Props = {
   title?: string;
 };
 
-const Thunks: NextPageWithLayout =  ({ title }: Props) => {
+const Confirm: NextPageWithLayout = ({ title }: Props) => {
   return (
     <>
       <Head>
         <title>{title}</title>
       </Head>
-      <p>THUNKS</p>
+      <p>CONFIRM</p>
     </>
   );
 };
 
-Thunks.getLayout = Page1.getLayout;
+Confirm.getLayout = Page1.getLayout;
 
-export default Thunks;
+export default Confirm;
