+++ src/component/pages/page1/page.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import type { NextPageWithLayout } from 'next';
 import Head from 'next/head';
-import { Page1Layout } from '@/components/layout';
+import { Page1Layout } from '@/component/layout';
 
 type Props = {
   title?: string;
@@ -21,4 +21,3 @@ const Page1: NextPageWithLayout = ({ title }: Props) => {
 Page1.getLayout = (page: ReactNode) => <Page1Layout>{page}</Page1Layout>;
 
 export default Page1;
-
