+++ src/pages/home/index.tsx
@@ -1,1 +1,1 @@
-export { default } from '@/components/pages/home/page';
+export { default } from '@/component/pages/home/page';
