+++ src/pages/page1/thunks/index.tsx
@@ -1,1 +1,1 @@
-export { default } from '@/components/pages/page1/thunks/page';
+export { default } from '@/component/pages/page1/thunks/page';
